requirements.txt
altair==5.3.0
attrs==23.2.0
azure-core==1.30.2
azure-storage-blob==12.20.0
blinker==1.8.2
cachetools==5.3.3
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cryptography==42.0.8
cycler==0.12.1
fonttools==4.53.0
gitdb==4.0.11
GitPython==3.1.43
humanize==4.10.0
idna==3.7
importlib_resources==6.4.0
isodate==0.6.1
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
mdurl==0.1.2
numpy==2.0.0
packaging==24.1
pandas==2.2.2
pillow==10.3.0
plotly==5.22.0
protobuf==5.27.2
pyarrow==16.1.0
pycparser==2.22
pydeck==0.9.1
Pygments==2.18.0
pyodbc==4.0.39
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2024.1
referencing==0.35.1
requests==2.32.3
rich==13.7.1
rpds-py==0.18.1
scikit-learn==1.5.0
scipy==1.13.1
six==1.16.0
smmap==5.0.1
SQLAlchemy==1.4.46
st-theme==1.2.3
streamlit==1.36.0
streamlit-autorefresh==1.0.1
streamlit-navigation-bar==3.3.0
tenacity==8.4.2
threadpoolctl==3.5.0
toml==0.10.2
toolz==0.12.1
tornado==6.4.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
watchdog==4.0.1
xgboost==2.1.1
zipp==3.19.2
